chore: move TS keywords to Node
