+++ packages/ast-spec/src/token/spec.ts
@@ -10,13 +10,4 @@ export * from './NumericToken/spec';
 export * from './PunctuatorToken/spec';
 export * from './RegularExpressionToken/spec';
 export * from './StringToken/spec';
-export * from './TSAbstractKeyword/spec';
-export * from './TSAsyncKeyword/spec';
-export * from './TSDeclareKeyword/spec';
-export * from './TSExportKeyword/spec';
-export * from './TSPrivateKeyword/spec';
-export * from './TSProtectedKeyword/spec';
-export * from './TSPublicKeyword/spec';
-export * from './TSReadonlyKeyword/spec';
-export * from './TSStaticKeyword/spec';
 export * from './TemplateToken/spec';
