+++ packages/ast-spec/src/type/TSAbstractKeyword/spec.ts
