+++ packages/ast-spec/src/type/TSDeclareKeyword/spec.ts
