+++ packages/ast-spec/src/type/TSPrivateKeyword/spec.ts
