+++ packages/ast-spec/src/type/TSProtectedKeyword/spec.ts
