+++ packages/ast-spec/src/type/TSReadonlyKeyword/spec.ts
