+++ packages/ast-spec/src/type/TSStaticKeyword/spec.ts
