+++ packages/ast-spec/src/type/spec.ts
@@ -1,9 +1,13 @@
+export * from './TSAbstractKeyword/spec';
 export * from './TSAnyKeyword/spec';
 export * from './TSArrayType/spec';
+export * from './TSAsyncKeyword/spec';
 export * from './TSBigIntKeyword/spec';
 export * from './TSBooleanKeyword/spec';
 export * from './TSConditionalType/spec';
 export * from './TSConstructorType/spec';
+export * from './TSDeclareKeyword/spec';
+export * from './TSExportKeyword/spec';
 export * from './TSFunctionType/spec';
 export * from './TSImportType/spec';
 export * from './TSIndexedAccessType/spec';
@@ -18,8 +22,13 @@ export * from './TSNullKeyword/spec';
 export * from './TSNumberKeyword/spec';
 export * from './TSObjectKeyword/spec';
 export * from './TSOptionalType/spec';
+export * from './TSPrivateKeyword/spec';
+export * from './TSProtectedKeyword/spec';
+export * from './TSPublicKeyword/spec';
 export * from './TSQualifiedName/spec';
+export * from './TSReadonlyKeyword/spec';
 export * from './TSRestType/spec';
+export * from './TSStaticKeyword/spec';
 export * from './TSStringKeyword/spec';
 export * from './TSSymbolKeyword/spec';
 export * from './TSTemplateLiteralType/spec';
