+++ packages/ast-spec/src/unions/TypeNode.ts
@@ -1,9 +1,13 @@
+import type { TSAbstractKeyword } from '../type/TSAbstractKeyword/spec';
 import type { TSAnyKeyword } from '../type/TSAnyKeyword/spec';
 import type { TSArrayType } from '../type/TSArrayType/spec';
+import type { TSAsyncKeyword } from '../type/TSAsyncKeyword/spec';
 import type { TSBigIntKeyword } from '../type/TSBigIntKeyword/spec';
 import type { TSBooleanKeyword } from '../type/TSBooleanKeyword/spec';
 import type { TSConditionalType } from '../type/TSConditionalType/spec';
 import type { TSConstructorType } from '../type/TSConstructorType/spec';
+import type { TSDeclareKeyword } from '../type/TSDeclareKeyword/spec';
+import type { TSExportKeyword } from '../type/TSExportKeyword/spec';
 import type { TSFunctionType } from '../type/TSFunctionType/spec';
 import type { TSImportType } from '../type/TSImportType/spec';
 import type { TSIndexedAccessType } from '../type/TSIndexedAccessType/spec';
@@ -18,7 +22,12 @@ import type { TSNullKeyword } from '../type/TSNullKeyword/spec';
 import type { TSNumberKeyword } from '../type/TSNumberKeyword/spec';
 import type { TSObjectKeyword } from '../type/TSObjectKeyword/spec';
 import type { TSOptionalType } from '../type/TSOptionalType/spec';
+import type { TSPrivateKeyword } from '../type/TSPrivateKeyword/spec';
+import type { TSProtectedKeyword } from '../type/TSProtectedKeyword/spec';
+import type { TSPublicKeyword } from '../type/TSPublicKeyword/spec';
+import type { TSReadonlyKeyword } from '../type/TSReadonlyKeyword/spec';
 import type { TSRestType } from '../type/TSRestType/spec';
+import type { TSStaticKeyword } from '../type/TSStaticKeyword/spec';
 import type { TSStringKeyword } from '../type/TSStringKeyword/spec';
 import type { TSSymbolKeyword } from '../type/TSSymbolKeyword/spec';
 import type { TSTemplateLiteralType } from '../type/TSTemplateLiteralType/spec';
@@ -35,12 +44,16 @@ import type { TSUnknownKeyword } from '../type/TSUnknownKeyword/spec';
 import type { TSVoidKeyword } from '../type/TSVoidKeyword/spec';
 
 export type TypeNode =
+  | TSAbstractKeyword
   | TSAnyKeyword
   | TSArrayType
+  | TSAsyncKeyword
   | TSBigIntKeyword
   | TSBooleanKeyword
   | TSConditionalType
   | TSConstructorType
+  | TSDeclareKeyword
+  | TSExportKeyword
   | TSFunctionType
   | TSImportType
   | TSIndexedAccessType
@@ -55,7 +68,12 @@ export type TypeNode =
   | TSNumberKeyword
   | TSObjectKeyword
   | TSOptionalType
+  | TSPrivateKeyword
+  | TSProtectedKeyword
+  | TSPublicKeyword
+  | TSReadonlyKeyword
   | TSRestType
+  | TSStaticKeyword
   | TSStringKeyword
   | TSSymbolKeyword
   | TSTemplateLiteralType
